app/requirements.txt
python-telegram-bot==21.4
requests>=2.31
reportlab>=4.0
pillow>=10.3
python-dotenv>=1.0
pytz
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
jinja2